+++ src/lightning_habana/pytorch/strategies/deepspeed.py
@@ -872,13 +872,13 @@ class HPUDeepSpeedStrategy(HPUParallelStrategy):
             )
         return False
 
-    def load_model_state_dict(self, checkpoint: Mapping[str, Any]) -> None:
+    def load_model_state_dict(self, checkpoint: Mapping[str, Any], strict: bool = True) -> None:
         # override to do nothing, deepspeed engine already loaded the weights in `load_checkpoint()`
         if self.load_full_weights and self.zero_stage_3:
             self.model_to_device()
-            self._restore_zero_state(checkpoint)
+            self._restore_zero_state(checkpoint, strict=strict)
 
-    def _restore_zero_state(self, ckpt: Mapping[str, Any]) -> None:
+    def _restore_zero_state(self, ckpt: Mapping[str, Any], strict: bool) -> None:
         """Overrides the normal load_state_dict behaviour in PyTorch.
 
         Ensure we gather parameters that may be sharded across processes before loading the
@@ -886,6 +886,7 @@ class HPUDeepSpeedStrategy(HPUParallelStrategy):
 
         Args:
             ckpt: The ckpt file.
+            strict: Flag to enable strict loading
 
         """
         import deepspeed
@@ -914,7 +915,7 @@ class HPUDeepSpeedStrategy(HPUParallelStrategy):
                         state_dict=state_dict,
                         prefix=prefix,
                         local_metadata=local_metadata,
-                        strict=True,
+                        strict=strict,
                         missing_keys=missing_keys,
                         unexpected_keys=unexpected_keys,
                         error_msgs=error_msgs,